changed wiimc version

+++ source/mplayer/osdep/shmem.h
@@ -19,9 +19,7 @@
 #ifndef MPLAYER_SHMEM_H
 #define MPLAYER_SHMEM_H
 
-#include <stdint.h>
-
-void* shmem_alloc(int64_t size);
-void shmem_free(void* p,int64_t size);
+void* shmem_alloc(int size);
+void shmem_free(void* p,int size);
 
 #endif /* MPLAYER_SHMEM_H */
